backend/requirements/base.txt
asgiref==3.7.2
black==23.9.1
boto3==1.28.46
botocore==1.31.46
click==8.1.7
Django==4.2.5
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
